test: 덱 셔플 기능 테스트 코드 구현
